+++ frontend/src/app/(no_auth)/login/page.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import { useRouter } from "next/navigation";
-import { ChangeEvent, FormEvent, useState } from "react";
+import { ChangeEvent, FormEvent, useEffect, useState } from "react";
 
 import { FeedbackModal } from "@/components/FeedbackModal";
 import { AuthPageShell } from "@/components/AuthPageShell";
@@ -14,9 +14,8 @@ const authFieldClassName =
 const authPrimaryButtonClassName =
   "h-10 sm:h-14 w-full rounded-md sm:rounded-lg bg-[var(--dark-blue)] text-sm sm:text-sm font-bold text-white transition-opacity hover:opacity-90";
 
-export default function Register() {
+export default function Login() {
   const router = useRouter();
-  const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [feedback, setFeedback] = useState<{
@@ -27,14 +26,14 @@ export default function Register() {
     variant: "success" | "error";
   } | null>(null);
 
-  const handleRegister = async (event: FormEvent<HTMLFormElement>) => {
+  const handleLogin = async (event: FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
-    if (!name || !email || !password) {
+    if (!email || !password) {
       setFeedback({
         open: true,
         title: "Missing details",
-        description: "Name, email, and password are required to register.",
+        description: "Email and password are required to sign in.",
         actionLabel: "Close",
         variant: "error",
       });
@@ -55,20 +54,20 @@ export default function Register() {
     }
 
     try {
-      const response = await fetch(`${backendUrl}/api/auth/register`, {
+      const response = await fetch(`${backendUrl}/api/auth/login`, {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
         },
-        body: JSON.stringify({ name, email, password }),
+        body: JSON.stringify({ email, password }),
       });
 
       if (!response.ok) {
         setFeedback({
           open: true,
-          title: "Registration failed",
+          title: "Sign in failed",
           description:
-            "We couldn't create your account. Please check your details and try again.",
+            "We couldn't sign you in with those credentials. Check them and try again.",
           actionLabel: "Try again",
           variant: "error",
         });
@@ -97,8 +96,8 @@ export default function Register() {
 
       setFeedback({
         open: true,
-        title: "Account created",
-        description: "Your account was created successfully. Redirecting...",
+        title: "Welcome back",
+        description: "You have successfully signed in. Redirecting...",
         actionLabel: "Continue",
         variant: "success",
       });
@@ -120,15 +119,15 @@ export default function Register() {
 
   return (
     <AuthPageShell
-      title="Register"
-      subtitle="Welcome to the club :)"
-      imageSrc="/9a1762f5a9668b3a87bb6b6f828eb28a8e10882c.jpg"
+      title="Sign In !!!"
+      subtitle="Welcome back :D"
+      imageSrc="/login-left.jpg"
       imageAlt="pixel room"
       footer={
         <p>
-          Have an account?{" "}
-          <Link href="/login" className="font-bold underline">
-            Log in here
+          Don’t have an account?{" "}
+          <Link href="/register" className="font-bold underline">
+            Register
           </Link>
         </p>
       }
@@ -146,18 +145,7 @@ export default function Register() {
         variant={feedback?.variant ?? "success"}
       />
 
-      <form onSubmit={handleRegister} className="space-y-4">
-        <input
-          type="name"
-          placeholder="Name"
-          value={name}
-          onChange={(e: ChangeEvent<HTMLInputElement>) =>
-            setName(e.target.value)
-          }
-          className={authFieldClassName}
-          required
-        />
-
+      <form onSubmit={handleLogin} className="space-y-4">
         <input
           type="email"
           placeholder="Email"
@@ -179,11 +167,18 @@ export default function Register() {
           className={authFieldClassName}
           required
         />
+
+        <button
+          type="button"
+          className="text-left text-sm font-semibold text-[#627a94]"
+        >
+          Forgot Password?
+        </button>
+
         <button type="submit" className={authPrimaryButtonClassName}>
-          Create Account
+          Sign in
         </button>
       </form>
     </AuthPageShell>
   );
 }
-
